replicate ckan elb s3 logs to cyber s3 bucket
As part of an incident action, there is a need to replicate the logs of the ckan public elb into the cyber s3 bucket for processing by splunk.

It should be noted that the replication rule is only enabled for the production stack as Cyber is only interested in that. Unfortunately, the way that terraform 0.11.x works is we can't set a count for an nested resource. Hence, the replication rule is created for all environments but only enabled for the production environment.

In addition, we set the version to enabled as replication requires that. This would not have effect on the size of the s3 bucket because AWS writes new log files with new timestamp as the part of the file name so each log files is unique.