Reduce Storage Time of S3 Database Backups in Integration
The govuk-integration-database-backups S3 Bucket has over 30TB of
data in it. As we do not need long term backups of Integration
databases this PR reduces the time they are stored for
(from 30 days to 3). This PR also disables Glacier Backups
in Integration. To do this, two new lifecycle rules have been added.
They will be created in all environments, but only enabled in Integration.