Update ASG scheduled action start/end dates
This commit updates the scheduled action start times for the Knowledge Graph ASG to ensure that we continue to generate base data at the weekend. This is necessary as we need an uninterrupted stream of data to take advantage of aggregating local data so that we don't have to query other data sources, in addition to providing source data for our data platform.