Increase MySQL RDS storage to recommended minimum and auto-scale.
Amazon recommends at least 100 GB of allocated storage when using
General Purpose IOPS, otherwise the I/O will be underprovisioned. Since
we're about to migrate Whitehall to this RDS deployment, we want to
avoid any potential performance issues.

Also enable autoscaling up to 800 GB (the same limit that we have for
the Postgres RDS instances). This reduces the risk of running out of
space and reduces unnecessary work in managing the storage allocation
manually.

Fix the `replica` instance so that it has the same allocated storage as
the primary. Previously this was erroneously set to 10 GB and would have
failed if the Terraform had been applied.