Give search-api permission to invoke SageMaker endpoints
The actual endpoint won't be managed by terraform so, rather than have
names in three places which need to be kept in sync (sagemaker
management code in concourse, search-api config, terraform), just
grant access to all endpoints.  We only have one at the moment, and
this can be revisited if need be.