Abort deploy script if an invalid project is given
The `deploy` script accepts four fields: `stack`, `project`,
`environment` and `command`. For each of them, there are only specific
possible values.

For the `stack`, `environment` and `command` fields, these values were
specified in the script and there was logic to abort if different values
were given.

This introduces the same logic for the `project` field.

I also reordered the variables and the `abort` statements to reflect the
order in which the fields must be given in the command line.