Add ability to monitor fastly_logs Athena databases
We want to have the means to monitor whether our fastly_logs athena
databases continue to return results, this is in response to a bad
deploy a few months ago where we discovered that the database was no
longer queryable.

To monitor Athena we need a user that has access to Athena, to AWS Glue
to access table details, to have read access to the bucket the data is
in and to have write access to a bucket to store results. The intention
is that this users credentials will be stored within Jenkins and
periodically a job will run that makes a query to Athena. This job will
be deemed successful if results can be pulled from the database.

This PR adds a new IAM user for this monitoring and creates a bucket
where results can be stored. The bucket has a short retention of 7 days
as I don't anticipate we'll have much (if any) use for the data it returns once
the query is completed. It then applies a policy so that this user can
access all the necessary resources.