Update docs using tools/update-docs.sh
In https://github.com/alphagov/govuk-aws/pull/1245 I'm seeing a failed
build due to docs not matching expected output. This failure seems to be
because terraform-docs has had a new release (0.8.2) rather the PR contents
so this is a separate commit to fix the issues we have with the new
version.

This has seemed to have broken the situations where we embed code with
three backticks.