Update Knowledge Graph to make use of central data generation
This commit updates the Knowledge Graph by decreasing the instance type and EBS requirements, as well as changing ASG scheduled actions to now only run the machine during the working week. It also updates the parameters passed into the provisioning script, removing those which are no longer necessary. Finally, outputs are added for the data infrastructure bucket name and ARN to read/write from/to this bucket by the central data generation instance (until these resources are migrated across to that project).