Move IAM policy for S3 assets sync to db_admin.
This is in order to migrate the assets sync job from using
alphagov/env-sync-and-backup on the Carrenza `asset_master` machine to a
simpler cron job on the `db_admin` machine (like all the other sync jobs
in AWS).

The plan is to run both the sync-to-staging and sync-to-integration jobs
from the Prod `db_admin` machine. This keeps the privileges minimal and
avoids the need for cross-account access into the Prod buckets.