Create an ASG for concourse to use for search LTR data generation
This commit creates an ASG, where the node template uses the same role
as the regular search machines, so members will be able to access
Elasticsearch and S3 just as search-api currently does.

Currently training data is generated by sending an HTTP request to
search-api and waiting.  But we want to now start producing more
training data, which will significantly slow that down.  We don't want
to put lots of extra strain on a live search-api instance, so instead
here's how it's going to work:

1. Concourse will scale up the ASG to 1 member
2. Concourse will SSH in to the new instance, then:
   1. Clone search-api from github
   2. Install dependencies
   3. Run the training data generation rake task
3. Concourse will scale down the ASG to 0 members