Update the scaling times of the Knowledge Graph ASG
This commit updates the scaling times of the Knowledge Graph auto scaling group to maintain the expected availability of the platform now that we are in UTC+1. Auto scaling group scheduled actions do not cater for any other timezone than UTC, so this is a necessary change to ensure the graph is available when people expect it to be.