Update the scaling times of the data science data ASG
This commit updates the scaling times of the data science data auto scaling group now that we are in UTC+1. Auto scaling group scheduled actions do not cater for any other timezone than UTC, so this is a necessary change to ensure the data generation happens consistently at the same time each day.