Update environment provisioning documentation
The documentation has been updated to reflect the additional
steps required to provision a new environment, and more detail
where that is helpful.

Bringing up a new environment is challenging. We're looking at
how we can make bringing up a new environment faster as part
of our discovery on the replatforming team.

Closes #1306.

Co-authored-by: Issy Long <issy.long@digital.cabinet-office.gov.uk>
Co-authored-by: Mahmud Hussain <mahmud.hussain@digital.cabinet-office.gov.uk>