Speed up cloning govuk-secrets to puppetmasters
The puppet masters clone govuk-secrets, because govuk-secrets is now so
large it takes a very long time to provision a puppetmaster.

This change only pulls down the git repo to a depth of 1 rather than
everything back to the triassic period.  This greatly increases the speed
in which to provision a puppetmaster.