Add app-accessibility-reports project
This commit adds the `app-accessibility-reports` project, which exists to run the numerous accessibility reports that are being created to help the Accessibility Team audit the GOV.UK content estate for WCAG compliance. The EC2 instance will run in the Tools account, accessing certain buckets in S3 in the integration and staging accounts through the `govuk-accessibility-reports-integration-data-reader` role, which has already been granted access in those accounts.