Update infra-mirror-bucket to output replica policy ARN
This commit adds two policies for cross account access to the govuk-mirror and govuk-mirror-replica buckets respectively, both granting access to the root ARN of the integration account. This is done so that the accessibility report runner can read mirror content when running on the integration account, via a role which grants it this access.