Remove sent and finished_sending_at from schema
Removes two columns from the email-alert-api archive schema. These are
`finished_sending_at_utc` and `sent`. The reason we are removing these
is they may change after an email has been archived, which
makes them misleading. For archiving purposes, it should be enough to
know when the email was created, since this is usually similar to when
it was sent.

Trello:
https://trello.com/c/15eVh9Qg/482-stop-relying-on-notify-status-updates-to-archive-delete-emails