Add 6 more Mapit instances
Unfortunately the Terraform config is currently hard coded for 2
instances, so to up it to 8, I've had to copy and paste the resources
manually.

It's definitely worth refactoring this in the future, but for now we
need to get the instances up relatively quickly, so I think this will do
for now.