Change scheduled action times for app-data-science-data and app-knowledge-graph
This commit updates the scheduled action times for spinning-up and spinning-down instances of both `app-data-science-data` and `app-knowledge-graph`. The reason for doing this is because data that we rely on in GA as part of the data generation pipeline is often not created until 08:30 UTC daily, so we should hold off until then to ensure that it's available.

The Knowledge Graph uses the output of the data generation pipeline, so it is also moved back an hour to ensure that output exists when the machine spins-up.