Set cache asg_min_size to asg_desired_capacity - 1
Having the minimum capacity be 1 less than the max/desired capacity enables
us to move one instance into a standby state, without the ASG starting a new
instance to replace the instance we are performing maintenance on.

This is helpful for rebooting an instance. We can follow the procedure detailed
here: https://docs.aws.amazon.com/autoscaling/ec2/userguide/as-enter-exit-standby.html#standby-state
which lets us reboot a machine safely (preventing it from serving requests while
it is rebooting).