Make LB actually check content store health
Previously the LB would simply check if Nginx was running on port
80. Since the Content Store is the only app running on machines of
the same name, we can more usefully check the app itself is healthy,
so that the LB will redirect traffic if it's not.

Note that the "/_healtcheck-<name>" route already exists on each of
the Content Store machines [1]. Most of our apps already use this
technique [2], either at the LB level, or the target group level (if
a machine has multiple apps to switch between).

[1]: https://github.com/alphagov/govuk-puppet/blob/00043c111f7939700539691472935c24c4a5838f/modules/govuk/manifests/app/config.pp#L304-L309
[2]: https://github.com/alphagov/govuk-aws/blob/5a2567dcc0f336c0fe61b4b3e277fb9059b9cf7b/terraform/projects/app-email-alert-api/main.tf#L102