Allow grafana to make requests to prometheus
Currently our grafana is configured with a prometheus data source that
expects to talk to prometheus on:

https://prometheus.production.govuk-internal.digital

but there's no DNS entry for `prometheus.production.govuk-internal.digital`,
so I don't think that can ever have worked.

If we wanted it to work, we'd need to create an internal load balancer
for prometheus, and then publish its address in route53.

That would be consistent, but it seems like a pain to me - do we really
want to run a separate load balancer just for this traffic?

Instead I suggest we let Grafana reach prometheus on the public URL
instead:

https://prometheus.production.govuk.digital

(note `govuk.digital`, not `govuk-internal.digital`)

It can't do this at the moment because prometheus's external load
balancer only allows requests from office IPs, and Grafana doesn't have
an office IP.

Adding this security group rule will allow requests from Grafana (which
is on the graphite VM) to Prometheus.