Add IAM user and policy for Ask Export
This creates an IAM User with an policy that allows permission to write
to an S3 bucket. This is needed as an export target for the GOV.UK Ask
service. Currently the module only uses the production backend (not
integration or staging) as cross account bucket permissions prevent use
in other environments.