Enable log_lock_waits PostgreSQL parameter
This setting allows us to look for information in the logs when our
queries are not able to execute because they are locked. The documentation
for the option is:

> Controls whether a log message is produced when a session waits longer
> than deadlock_timeout to acquire a lock. This is useful in determining
> if lock waits are causing poor performance. The default is off. Only
> superusers can change this setting.

I'd like to enable this to help with some investigation I'm doing
currently on some occasional timeouts we see in Email Alert API. I've
been looking in the RDS log files and for a specific time around 16:00
yesterday we have simple enough queries such as `SELECT "subscribers".*
FROM "subscribers" WHERE "subscribers"."id" = $1 LIMIT $2 FOR UPDATE`
taking over 10 minutes. I've identified that this is likely down to some
sort of locking because there are 8 similar queries which all complete
at exactly 16:11:48, which happens to coincide with a `COMMIT` from
a transaction which started at 15:57:58. Suspiciously, these times seem
to also coincide quite nicely with a bulk email sending task that was run
yesterday: https://deploy.blue.production.govuk.digital/job/send-bulk-email/9/