Set deadlock_timeout PostgreSQL parameter
This setting is the timeout which is used by PostgreSQL to detect
deadlocks when running queries. The documentation for this option is:

> This is the amount of time, in milliseconds, to wait on a lock before
> checking to see if there is a deadlock condition. The check for deadlock
> is relatively expensive, so the server doesn't run it every time it waits
> for a lock. We optimistically assume that deadlocks are not common in
> production applications and just wait on the lock for a while before
> checking for a deadlock. Increasing this value reduces the amount of time
> wasted in needless deadlock checks, but slows down reporting of real
> deadlock errors. The default is one second (1s), which is probably about
> the smallest value you would want in practice. On a heavily loaded server
> you might want to raise it. Ideally the setting should exceed your typical
> transaction time, so as to improve the odds that a lock will be released
> before the waiter decides to check for deadlock. Only superusers can
> change this setting.

This change might be a more contentious, or less scientific.

The main motivation behind this is to increase the time before we get
logs about queries waiting for locks (see
5af25cdb8a99ae5873b21c823300cd3e5ead883f). By increasing this threshold
I'm hoping we'll avoid getting lots of noise in our logs about queries
which are only waiting for a second whereas we're more interested (at
the moment) in the queries which seem to be waiting for much longer.

After reading the documentation for `deadlock_timeout` I started to
think that the default of 1 second is perhaps quite small for the sorts
of query durations we often get (especially Email Alert API and
Publishing API). The documentation says this setting should exceed our
typical transaction time, and although I have no doubt that our average
transaction time is well below 1 second, we often have transactions
which can take a long time to run. Checking for deadlocks is relatively
expensive so increasing this setting will reduce the potentially wasted
time spent looking for deadlocks in our long running queries.

We experience deadlocks pretty infrequently (looking through our most
recent logs I couldn't find an example of one being found) so I don't
feel like there are any major downsides to increasing the timeout from 1
second to 2.5 seconds.