Add lifecycle rule for lambdas prefix in data bucket
This commit adds a lifecycle rule for the `lambdas` prefix of the `govuk-data-infrastructure-integration` bucket in the integration account. The lifecycle rule will delete all previous versions of files under this prefix 30 days after they are made non-current versions, which will happen as new versions of Lamda deployment packages are uploaded.