Scale Mapit down to 4 instances
Mapit previously ran with 2 instances in production, prior to the
Coronavirus Restrictions Tier checker being released. As that is not
being used, we can scale the instances back.

As we are also upgrading Mapit to a new version, we need to stop Mapit
on one instance at a time to upgrade the application and database. We
will therefore retain 4 instances until this upgrade is complete, after
which we will reduce to 2 instances.