Deploy the ArtefactSync lambda from source, not a checked-in zip.
The ArtefactSync lambda broke in #1441 because the zip wasn't updated,
which is completely understandable because it's a surprising and
unnecessary manual process which was neither discoverable nor
documented.

Instead of relying on an undocumented manual process of zipping up
main.py and checking it into the repo before running Terraform, have
Terraform build the zip itself so that this breakage can't happen again.

It'd be even better to put the zip in S3 by defining an s3_bucket_object
resource for it, but I'll leave that as a possible future improvement.
There's precedent for this half-way-house solution of having Terraform
build and upload the zip directly: the WAFLogTrimmer lambda
(infra-public-services/waf.tf) is currently set up this way.

The TransitionLogs and CloudfrontUrlRewrite lambdas have the same
problem, but this commit only addresses ArtefactSync.