Output role names and ARNs for roles created via gds_user_role module
This commit updates the `gds_user_role` module to output a map containing the role name and role ARN for each created role.

For example, creating a role of `*-admin` will output all of the generated role names and ARNs in a map of the format:

```
{
  ...,

  'joe.bloggs-admin': 'arn:aws:iam::123467890123:role/joe.bloggs-admin',

  ...,
}
```

This is being used in `infra-security` to output the role names and ARNs when creating `admin`, `poweruser` and `user` roles, allowing us to use this when creating resources elsewhere, such as in our new EKS cluster.

Co-Authored-By: Chris Banks <chris.banks@digital.cabinet-office.gov.uk>