Automatically replace unhealthy Apt instances
This changes the health check type to ELB for the Apt
autoscaling group, and increases the asg_max_size to 3.

The effect of this will be that apt instances will be
deregistered from the ASG and replaced when their ELB
health checks fail.

I have also increased the grace period for a new instance
to 5 minutes as this instance may take a while to start.