Set a two-year expiration time on the GOV.UK accounts backup
We've taken a snapshot of the account-manager[1] and
attribute-service[2] databases before terminating the apps.  This is
in case we need to refer to any historic data (most likely the audit
log in the security_activities table, which didn't get migrated to
Digital Identity). We've said in our privacy policy that we keep user
data for two years, so we should get rid of the data at that point.

[1] https://github.com/alphagov/govuk-account-manager-prototype
[2] https://github.com/alphagov/govuk-attribute-service-prototype