Remove migration-hosts.sh
This shell script is failing shellcheck, causing CI to fail in unrelated
branches (https://github.com/alphagov/govuk-aws/runs/4445343066).

The script looks like a hold over from the Carrenza->AWS migration, so
we can just remove it rather than fixing it. Also removing a bit of
documentation that refers to the script.