Configure postgresql-primary for AWS DMS
This changes the configuration of the postgresql-primary machine to be
set-up so that it can be a source for AWS's Database Migration Service.

This configuration is taken from Amazon's documentation [1] and the
experience I had of errors when pglogical wasn't an available library.

Once we have completed DMS work these parameters can be removed,
although if everything goes to plan this instance will also be removed
then.

[1] https://docs.aws.amazon.com/dms/latest/userguide/CHAP_Source.PostgreSQL.html#CHAP_Source.PostgreSQL.RDSPostgreSQL.CDC