Configure mysql-primary for AWS DMS
This changes the configuration of the mysql-primary machine to be
set-up so that it can be a source for AWS's Database Migration Service. This
configuration is taken from Amazon's documentation [1]. These are all
dynamic parameters so they do not require a reboot to apply.

This will also require an additional manual step of running:

```
 call mysql.rds_set_configuration('binlog retention hours', 24);
```

on each instance. I'm not aware of a way this can be done via Terraform
configuration.

Once we have completed DMS work these parameters can be removed,
although if everything goes to plan this instance will also be removed
then.

[1]: https://docs.aws.amazon.com/dms/latest/userguide/CHAP_Source.MySQL.html#CHAP_Source.MySQL.AmazonManaged