Add internal DNS records for RDS instances
This came from app-postgresql and app-mysql, which are available at
`{mysql,postgresql}-primary.${internal_domain_name}`.

For example, the new account-api instance will be available at:
`account-api-postgresql.blue.integration.govuk-internal.digital`.

I'm not sure if this is useful, but I see no reason to not be
consistent with what came before.