Create one db-admin instance per RDS instance
This is a large commit which, unfortunately, can't really be sensibly
broken up into smaller chunks.  It's essentially the `node_group`
module, specialised to the parameters the `app-db-admin` project uses.

It feels a bit overkill creating an ASG for a single instance, but I
think this is the better approach.  Just creating an EC2 instance with
Terraform directly would work, but if we ever terminated the
instance (eg, it had an issue we didn't want to invest time into
fixing, and just wanted to restore a clean state), we'd have to re-run
terraform to create a new instance.  Using the ASG, while rather more
verbose, avoids this.

An instance will be assigned the Puppet node class
`s_${database}_db_admin`: this comes from the `aws_migration` tag on
the ASG (which instances inherit), and Puppet adding the `s_` prefix.
See the govuk-puppet docs[1].

Instances are available at `${database}-db-admin.${internal_domain_name}`

One slight deviation from usual conventions I've had to make is in
naming the ELBs. I've called them `rds-${database}`, rather than
something like `blue-govuk-rds-${database}`, because of a length
restriction.

[1] https://github.com/alphagov/govuk-puppet/blob/main/docs/puppet-class-assigment.md