Simplify lifecycle rules for database backups
This commit simplifies the lifecycle rules on the S3 buckets which hold database backups created by env_sync.

Previously, various retention rules were needed on different paths within the bucket. This reflected the fact that the bucket was being used to store lots of different types of backup.

We made the decision to simplify the purpose of this bucket so that it's only used to store env_sync database dumps.

We now have one lifecycle rule which is applied to the entire bucket.

Given the current configuration, this means:
- Production backups will be permanently deleted after 121 days ('soft deleted' after 120 days and 'hard deleted' the next day)
- Staging & integration backups will be permanently deleted after 4 days ('soft deleted' after 3 days)

This database backups bucket (and its replica bucket) have versioning enabled. We're dubious of the value this provides – it seems to complicate the lifecycle rules, and only protects against the unlikely case of accidental deletion. However after much deliberation, we decided not to turn it off (that was one 'yak shave' too far).

We also recognise that the lifecycle rules are duplicated between buckets, and so could probably be DRY'd up. Again, this was one 'yak shave' too far. Our intention is simply to leave this configuration in a better state than it was before.