Enable ELB health checks on calc_frontend instances
This will mean that both EC2 and ELB checks are used
to determine whether an instance of a calculators_frontend
is healthy or should be removed from an ASG and terminated.

The result is that an app failing a /healthcheck/ready check
on an instance will lead to the instance being removed from an
AutoScalingGroup.

As far as I can see until now we have needed to terminate
instances manually when application level checks had failed.

I'm starting with calculators_frontend and will roll this out to
other instances in future.

Docs: https://docs.aws.amazon.com/autoscaling/ec2/userguide/as-add-elb-healthcheck.html