Add shield protection to public resources
This has already been enabled for some resources in production, so this is
a bit of a paperwork catch up in a lot of cases.

Enabling shield protection allows us to add other features later.

I've tweaked the name of the existing one, as I want it to refer to the load
balancer rather than the web acl.

https://trello.com/c/qhzXqw3I/2870-ensure-that-resources-added-to-aws-shield-advanced-are-terraformed